src/training/requirements.txt
compress_pickle==2.1.0
datasets==2.19.2
flwr==1.12.0
flwr_datasets==0.4.0
py_solc_x==2.0.3
web3==7.3.0
xgboost==2.1.2
